Pipelining the describe function:  - Add dask-based compute graph solution  - Split some functions to make the data flow clear  - Use a unified progress bar
